divide thread_safe_data
